pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["python"]

[project]
name = "spws-ranklist"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "rapidfuzz>=3.9",
    "openpyxl>=3.1",
    "xlrd>=2.0",
    "odfpy>=1.4",
    "supabase>=2.4",
    "pypdf>=4.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "respx>=0.21",
    "psycopg2-binary>=2.9",
]

[tool.pytest.ini_options]
testpaths = ["python/tests"]
asyncio_mode = "auto"
markers = [
    "integration: tests requiring live external APIs (deselect with -m 'not integration')",
]
